read microcontroller data. WIP
Updating the read file method and then the GUI to reflect the new read data method. As opposed to parsing through each thing from the GPS (lat, long, etc), have the microntroller send a single bit that corresponds to a city, then by reading in that bit we display the correct city. Still a work in progress